refactor: turnManager 삭제, 턴 변경 로직 위치 이동
